Consolidate container session stuff
